Moved TestEntiy to its own package
